grafana/ui: Remove Stack and rename FLex to Stack
