+++ packages/grafana-ui/src/components/Layout/Box/Box.mdx
@@ -16,7 +16,7 @@ Use it whenever you would use custom CSS.
 
 #### When not to use
 
-If you need layout styles, use the Stack, Flex or Grid components instead.
+If you need layout styles, use the Stack or Grid components instead.
 
 ### How to add a prop to Box
