+++ packages/grafana-ui/src/components/Layout/Grid/Grid.mdx
@@ -17,11 +17,6 @@ Use the Grid component when you want to create structured and organized layouts 
 
 Use the `Stack` component instead for these use cases:
 
-- **Simple layouts:** When you need to arrange elements in a linear format, either vertically or horizontally.
-- **Regular flow:** When you want a "regular" site flow but with standardized spacing between the elements.
-
-Use the `Flex` component instead for these use cases:
-
 - **Alignment:** More options for item alignment.
 - **Flex items:** Custom flex basis or configure how items stretch and wrap.
