--- packages/grafana-ui/src/components/Layout/Stack/index.ts
@@ -1,2 +1,0 @@
-export { Stack } from './Stack';
-export { HorizontalStack } from './HorizontalStack';
